requirements.txt
Django==5.1.6
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-push-notifications==3.0.0
channels==4.0.0
daphne==4.0.0
asgiref==3.9.2
gunicorn==21.2.0
whitenoise==6.6.0
python-dotenv==1.0.0
Pillow==10.4.0
boto3==1.34.0
django-storages==1.14.2